Signout-add

+++ src/actions/auth.actions.js
@@ -42,22 +42,22 @@ export const signup = (user) => {
         dispatch({ type: authConstants.LOGIN_REQUEST });
         const res = await axios.post(`/signup`, {
             ...user,
-            
+
         })
-        
+
 
         if (res.status === 201) {
-            const {  token , message} = res.data;
+            const { token, message } = res.data;
             console.log(user, message)
             dispatch({
                 type: authConstants.LOGIN_SUCCESS,
                 payload: {
-                    token, message,  user
+                    token, message, user
                 }
             });
         }
         else {
-            console.log('else',user);
+            console.log('else', user);
             if (res.status === 400) {
                 dispatch({
                     type: authConstants.LOGIN_FAILURE,
@@ -70,10 +70,10 @@ export const signup = (user) => {
     }
 }
 
- export const isUserLoggedIn = () =>{
-    return async dispatch  => {
+export const isUserLoggedIn = () => {
+    return async dispatch => {
         const token = localStorage.getItem('token');
-        if(token){
+        if (token) {
             const user = JSON.parse(localStorage.getItem('user'));
             dispatch({
                 type: authConstants.LOGIN_SUCCESS,
@@ -81,7 +81,7 @@ export const signup = (user) => {
                     token, user
                 }
             });
-        }else{
+        } else {
             dispatch({
                 type: authConstants.LOGIN_FAILURE,
                 payload: { error: 'Failet to login ' }
@@ -90,11 +90,25 @@ export const signup = (user) => {
     }
 }
 
-export const signout = () =>{
-    return async dispatch =>{
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        });
+export const signout = () => {
+    return async dispatch => {
+
+        dispatch({type : authConstants.LOGOUT_REQUEST});      
+
+        const res = await axios.post('/signout');
+        console.log(`------------------${res}`)
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({type: authConstants.LOGOUT_SUCCESS});
+        }
+        else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: {error : res.data.error}
+            })
+        }
+
+
+
     }
 }
